requirements.txt
# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg[binary]>=3.1.0
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# APIs e Comunicação
requests>=2.28.0
websocket-client>=1.4.0
httpx>=0.25.0
aiohttp>=3.8.0

# Análise de Dados
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0
ta>=0.10.2

# Processamento de Linguagem Natural
nltk>=3.7.0

# Visualização e Dashboard
dash>=2.6.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0
plotly>=5.10.0

# Armazenamento de Dados
psycopg2-binary>=2.9.5

# Utilidades
python-dotenv>=0.21.0
psutil>=5.9.0
certifi>=2023.0.0
typing-extensions>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
